Add BFF project and update solution configurations
Updated the Visual Studio version in the solution file to 17.14.36301.6 and added a new "BFF" project with necessary configurations. Enhanced Docker Compose files to include the "BFF" service, enabling it to run alongside "mystore.api". Improved logging settings in appsettings files and added HTTP client configuration in Program.cs for API interactions. Developed a Dockerfile for the "BFF" service to streamline image creation and deployment.